chore: restructure repo following golang-standards
