+++ .dockerignore
@@ -1,0 +1,5 @@
+CODEOWNERS
+LICENSE
+Makefile
+README.md
+deploy
