+++ Makefile
@@ -4,28 +4,33 @@ TARGET_OS ?= linux
 # When developing locally, change this to whatever fqdn you are using for 127.0.0.1
 DOMAIN ?= localhost
 
+COMPOSE_OPTS := -f deploy/docker-compose.yml
+DOCKER_OPS := -f deploy/Dockerfile
 
 test:
 	go test ./... -v
 
+image:
+	docker build -t algolia/supersecretmessage $(DOCKER_OPS) .
+
 build: 
-	@docker-compose build
+	@docker-compose $(COMPOSE_OPTS) build
 
 clean:
-	@docker-compose rm -fv
+	@docker-compose $(COMPOSE_OPTS) rm -fv
 
 run-local: clean
         @DOMAIN=$(DOMAIN) \
-        docker-compose up --build -d
+	docker-compose $(COMPOSE_OPTS) up --build -d
 
 run: 
 	@DOMAIN=$(DOMAIN) \
-        docker-compose up --build -d
+        docker-compose $(COMPOSE_OPTS) up --build -d
 
 logs:
-	@docker-compose logs -f
+	@docker-compose $(COMPOSE_OPTS) logs -f
 
 stop:
-	@docker-compose stop
+	@docker-compose $(COMPOSE_OPTS) stop
 
-.PHONY: test build clean run-local run logs stop
+.PHONY: test image build clean run-local run logs stop
