+++ deploy/Dockerfile
@@ -2,10 +2,9 @@ FROM golang:1.18 AS builder
 
 WORKDIR /go/src/github.com/algolia/sup3rS3cretMes5age
 
-ADD . .
-
+COPY . .
 RUN go mod download
-RUN CGO_ENABLED=0 GOOS=linux go build -o sup3rS3cretMes5age .
+RUN CGO_ENABLED=0 GOOS=linux go build -trimpath -buildmode=pie -a -ldflags '-s -w -extldflags "-static"' -o /tmp/sup3rS3cretMes5age cmd/sup3rS3cretMes5age/main.go
 
 
 FROM alpine:latest
@@ -27,7 +26,7 @@ RUN \
 
 WORKDIR /opt/supersecret
 
-COPY --from=builder /go/src/github.com/algolia/sup3rS3cretMes5age/sup3rS3cretMes5age .
-COPY static /opt/supersecret/static
+COPY --from=builder /tmp/sup3rS3cretMes5age .
+COPY web/static/ /opt/supersecret/static/
 
 CMD [ "./sup3rS3cretMes5age" ]
