+++ web/static/application.css
