+++ web/static/icons/browserconfig.xml
